refactor models.Filter and models.Subtitle
