+++ bgmi/sql.py
@@ -10,7 +10,6 @@ CREATE_TABLE_BANGUMI = '''CREATE TABLE IF NOT EXISTS bangumi (
           status INTEGER DEFAULT 0
         )'''
 
-
 CREATE_TABLE_FOLLOWED = '''CREATE TABLE IF NOT EXISTS followed (
           id INTEGER PRIMARY KEY AUTOINCREMENT,
           bangumi_name TEXT NOT NULL UNIQUE,
@@ -18,7 +17,6 @@ CREATE_TABLE_FOLLOWED = '''CREATE TABLE IF NOT EXISTS followed (
           status INTEGER DEFAULT 1,
           updated_time INTEGER DEFAULT 0
         )'''
-
 
 CREATE_TABLE_DOWNLOAD = '''CREATE TABLE IF NOT EXISTS download (
           id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -29,7 +27,6 @@ CREATE_TABLE_DOWNLOAD = '''CREATE TABLE IF NOT EXISTS download (
           status INTEGER DEFAULT 0
         )'''
 
-
 CREATE_TABLE_FOLLOWED_FILTER = '''CREATE TABLE IF NOT EXISTS filter (
           id INTEGER PRIMARY KEY  AUTOINCREMENT,
           bangumi_name TEXT  UNIQUE NOT NULL,
@@ -39,12 +36,10 @@ CREATE_TABLE_FOLLOWED_FILTER = '''CREATE TABLE IF NOT EXISTS filter (
           regex TEXT
         )'''
 
-
 CREATE_TABLE_SUBTITLE = '''CREATE TABLE IF NOT EXISTS subtitle (
           id TEXT PRIMARY KEY UNIQUE NOT NULL,
           name TEXT
         )'''
-
 
 CREATE_TABLE_SCRIPT = '''CREATE TABLE IF NOT EXISTS scripts (
           id INTEGER PRIMARY KEY  AUTOINCREMENT,
